refactor: component for blogList
